Back-End-Rent-a-Ton/requirements.txt
asgiref==3.2.10
click==7.1.2
Django==3.0.7
django-cors-headers==3.4.0
django-postgres-dropdb==0.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pkg-resources==0.0.0
progressbar2==3.51.4
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
